@@ -1,28 +1,46 @@
+import java.text.SimpleDateFormat
+
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.xmut.el"
-        minSdkVersion 16
-        targetSdkVersion 28
+        minSdkVersion 22
+        targetSdkVersion 29
         versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        versionName "0.0.1"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+}
+android.applicationVariants.all { variant ->
+    variant.outputs.all { //修改apk名
+        if (variant.name == "release"){
+            outputFileName = "元素实验室_V${android.defaultConfig.versionName}_${variant.name}.apk"
+            println outputFileName
+        }else{
+            outputFileName = "元素实验室_V${android.defaultConfig.versionName}_${getCurrentDate()}_${variant.name}.apk"
+            println outputFileName
         }
     }
 }
 
+def getCurrentDate() {
+    SimpleDateFormat sdf = new SimpleDateFormat("MM-dd-HH-mm")
+    return sdf.format(new Date())
+}
+
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
 }
